Get settings from the CSV headers where possible
Adds lots of user input prompts to determine which columns from input
CSVs should be taken for the IDP ID and event PID.
Allows an input IDP ID to be specified in case the input data does
not include it.
Includes logic to try and reuse settings between files if possible,
though this is not utilised in the current hasher implementation.